+++ src/com/floreysoft/jmte/renderer/NullRenderer.java
@@ -1,0 +1,16 @@
+package com.floreysoft.jmte.renderer;
+
+import com.floreysoft.jmte.NamedRenderer;
+import com.floreysoft.jmte.Renderer;
+
+import java.util.Locale;
+import java.util.Map;
+
+/**
+ * Marker interface for a renderer that can render <code>null</code>.
+ * 
+ * @see NamedRenderer
+ * @see Renderer
+ */
+public interface NullRenderer {
+}